+++ pathtrace/sphere.h
@@ -5,10 +5,10 @@
 namespace pathtracer {
     struct Sphere : public Shape {
     public:
-        glm::vec3 position;
-        float radius;
+        glm::dvec3 position;
+        double radius;
 
-        std::optional<float> intersect(const Ray& ray) const override;
-        glm::vec3 normal(const glm::vec3& intersection) const override;
+        std::optional<double> intersect(const Ray& ray) const override;
+        glm::dvec3 normal(const glm::dvec3& intersection) const override;
     };
 }  // namespace pathtracer
